consistency fixes, updated readme, removed link storage if not specified
