+++ lib/exposure/exposure.js
@@ -42,6 +42,8 @@ export default class Exposure {
         const collection = this.collection;
         const firewall = this.firewall;
 
+        collection.__isExposedForGrapher = true;
+
         if (firewall) {
             collection.firewall = (filters, options, userId) => {
                 if (userId !== undefined) {
