+++ lib/links/linker.js
@@ -100,6 +100,10 @@ export default class Linker {
      */
     isMeta()
     {
+        if (this.isVirtual()) {
+            return this.linkConfig.relatedLinker.isMeta();
+        }
+
         return !!this.linkConfig.metadata;
     }
 
@@ -108,6 +112,10 @@ export default class Linker {
      */
     isSingle()
     {
+        if (this.isVirtual()) {
+            return this.linkConfig.relatedLinker.isSingle();
+        }
+
         return _.contains(this.oneTypes, this.linkConfig.type);
     }
