+++ lib/query/hypernova/assembler.js
@@ -8,22 +8,25 @@ export default (childCollectionNode, limit) => {
     const strategy = linker.strategy;
     const isVirtual = linker.isVirtual();
     const isSingle = linker.isSingle();
-    const oneResult = (isVirtual && linker.linkConfig.relatedLinker.linkConfig.unique)
-            || (!isVirtual) && isSingle;
+    const removeStorageField = !childCollectionNode.parentHasMyLinkStorageFieldSpecified();
+    const oneResult = (isVirtual && linker.linkConfig.relatedLinker.linkConfig.unique) || (!isVirtual) && isSingle;
 
     const fieldStorage = linker.linkStorageField;
 
     _.each(parent.results, result => {
-        result[childCollectionNode.linkName] = assembleData(childCollectionNode, result, {
-            fieldStorage, strategy, isVirtual, isSingle, oneResult, limit
+        const data = assembleData(childCollectionNode, result, {
+            fieldStorage, strategy, isVirtual, isSingle
         });
+
+        result[childCollectionNode.linkName] = filterAssembledData(data, {limit, oneResult})
     });
+
+    if (removeStorageField) {
+        _.each(parent.results, result => delete result[fieldStorage]);
+    }
 }
 
-function assembleData(childCollectionNode, result, {fieldStorage, strategy, isVirtual, oneResult, limit}) {
-    const filters = createSearchFilters(result, fieldStorage, strategy, isVirtual);
-    const data = sift(filters, childCollectionNode.results);
-
+function filterAssembledData(data, {limit, oneResult}) {
     if (limit) {
         return data.slice(limit);
     }
@@ -33,4 +36,10 @@ function assembleData(childCollectionNode, result, {fieldStorage, strategy, isVi
     }
 
     return data;
+}
+
+function assembleData(childCollectionNode, result, {fieldStorage, strategy, isVirtual}) {
+    const filters = createSearchFilters(result, fieldStorage, strategy, isVirtual);
+
+    return sift(filters, childCollectionNode.results);
 }
