+++ lib/query/nodes/collectionNode.js
@@ -19,10 +19,6 @@ export default class CollectionNode {
         return _.filter(this.nodes, n => n instanceof FieldNode);
     }
 
-    hasGlobalFieldNode() {
-        return !!_.find(this.fieldNodes, n => n.isGlobal());
-    }
-
     /**
      * @param node
      * @param linker
@@ -40,10 +36,6 @@ export default class CollectionNode {
      */
     applyFields(filters, options) {
         let hasAddedAnyField = false;
-
-        if (this.hasGlobalFieldNode()) {
-            return options.fields = undefined;
-        }
 
         _.each(this.fieldNodes, n => {
             hasAddedAnyField = true;
@@ -69,5 +61,17 @@ export default class CollectionNode {
         if (!hasAddedAnyField) {
             options.fields = {_id: 1};
         }
+    }
+
+    parentHasMyLinkStorageFieldSpecified() {
+        const storageField = this.linker.linkStorageField;
+
+        if (this.parent) {
+            return !!_.find(this.parent.fieldNodes, fieldNode => {
+                return fieldNode.name == storageField
+            })
+        }
+
+        return false;
     }
 }
