+++ lib/query/nodes/fieldNode.js
@@ -1,11 +1,7 @@
 export default class FieldNode {
     constructor(name, body) {
         this.name = name;
-        this.body = body;
-    }
-
-    isGlobal() {
-        return this.name === '$all';
+        this.body = _.isObject(body) ? 1 : body;
     }
 
     applyFields(fields) {
